Initial setup with UE4 and GitHub Desktop
feat: Added in the base of the application

Turn corners with the Oculus DK2 is now supported. Much of the
mechanices are working. Features left to do are: 1) Game Over screen
with distance and Faith points score 2)Growing glow of character to
indicate Faith points level 3)Calibrating the game difficulty based on
how the game feels and play now 4)Might include the flashlight on/off to
simulate limited battery life and delay for recharging