feat: Welcome and Game Over Screen + Fixes
So now we have a Welcome screen and a prologue which players can view if
they choose to. Input ideally should be from a gamepad. A Game Over
screen details the distance and faith points result.
Fixes were made to coordinate head turns during corner turns floortiles
though there is headshakes when PawnControlRotation is set to false to
coordinate these VR headturns  when in the Corner turn floortiles. It
used to be that when you died in the Corner turn floortile, you can no
longer use the HMD and turn and look around. This is fixed by setting
the PawnControlRotation to True again for the FPCamera before the Death
function is called.
To Dos: Calibrate the Game difficulty. Implement switching on and off
the flashlight and double the faith points collected when light is off.
Would-be-nice-to-fix: Look into better lighting especially the Forest.
Implement glowing effect of the player character to simulate growth in
faith when collecting points. Faint blue glow when collecting blue
faith, faint red glow when collecting red faith.