fix: Amit's audio work is now integrated
There are over a 100 changes because I used Amit's copy of the project
which has his audio work done. I have made a lot of fixes and
improvements to my own side of things too. Added in difficulty
calibration based on distance travelled and the amount of faith points
the player has attained. Also I fixed the brightness of the spotlight,
this is dependent on the property "Use Inverse Squared Falloff" which
when enabled closely replicates a very realistic behaviour  of light in
the real world It causes light to be very bright when clsoet to its
ources, and dims very quickly when it moves away. Disabling it allows
the spotlight to be like a bright beacon, and become really bright and
strong.
To fix: The jumping into portal when the base desired direction and
forest desired direction are 180 degrees apart still causes the player
to fall off the map because of the air velocity and momentum. Still have
yet to fix this.