feat: Replaced the controls to turn at corners with the left and right trigger
The use of the left thumbstick is for strafing the player left and right
when it is constantly running forward. This somehow confuses the player
to use it too for turning at corners therefore we have now seperated it
for using the left thumbstick to strafe and triggers to turn at corners.